Rename Python `graph` to `framework`
