+++ python/paddle/v2/framework/tests/test_program.py
@@ -1,7 +1,7 @@
 import unittest
 
 import paddle.v2.framework.core as core
-from paddle.v2.framework.graph import g_program
+from paddle.v2.framework.framework import g_program
 
 
 class TestProgram(unittest.TestCase):
